backend/requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis and Celery
redis==5.0.1
celery==5.3.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation and Settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# WebSocket
websockets==12.0

# PDF Generation
weasyprint==60.2

# Utilities
python-dateutil==2.8.2
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
